+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/DoEverythingReqRepThriftMethodBuilderClientModuleFeatureTest.scala
@@ -1,70 +1,60 @@
 package com.twitter.inject.thrift
 
+import com.twitter.conversions.time._
 import com.twitter.finagle.http.Status.Ok
 import com.twitter.finatra.http.{EmbeddedHttpServer, HttpTest}
 import com.twitter.finatra.thrift.EmbeddedThriftServer
 import com.twitter.greeter.thriftscala.Greeter
 import com.twitter.inject.server.FeatureTest
-import com.twitter.inject.thrift.integration.serviceperendpoint.{
-  EchoServicePerEndpointModule,
-  EchoThriftService,
-  GreeterServicePerEndpointModule,
-  GreeterThriftService,
-  ServicePerEndpointHttpController
-}
+import com.twitter.inject.thrift.DoEverythingReqRepThriftMethodBuilderClientModuleFeatureTest._
+import com.twitter.inject.thrift.integration.reqrepserviceperendpoint.{ReqRepServicePerEndpointHttpController, GreeterReqRepThriftMethodBuilderClientModule, GreeterThriftService}
 import com.twitter.inject.thrift.integration.{TestHttpServer, TestThriftServer}
-import com.twitter.test.thriftscala.EchoService
+import com.twitter.util.Duration
+import com.twitter.util.tunable.Tunable
 
-class DoEverythingServicePerEndpointModuleFeatureTest extends FeatureTest with HttpTest {
+object DoEverythingReqRepThriftMethodBuilderClientModuleFeatureTest {
+  case class HelloHeaders(empty: Boolean)
+  case class HelloResponse(value: String, headers: HelloHeaders)
+}
+
+class DoEverythingReqRepThriftMethodBuilderClientModuleFeatureTest
+  extends FeatureTest
+  with HttpTest {
   override val printStats = false
 
+  private val requestHeaderKey = "com.twitter.greeter.test.header"
   private val httpServiceClientId = "http-service"
+  private val perRequestTimeoutTunable: Tunable[Duration] = Tunable.mutable("per-request", 50.millis)
 
   private val greeterThriftServer = new EmbeddedThriftServer(
-    twitterServer =
-      new TestThriftServer(new GreeterThriftService(httpServiceClientId).toThriftService)
+    twitterServer = new TestThriftServer(
+      new GreeterThriftService(
+        httpServiceClientId,
+        requestHeaderKey).toThriftService)
   )
 
-  private val echoThriftServer = new EmbeddedThriftServer(
-    twitterServer = new TestThriftServer(new EchoThriftService(httpServiceClientId).toThriftService)
-  )
-
-  // Test an HttpServer with multiple clients
-  val server = new EmbeddedHttpServer(
-    twitterServer = new TestHttpServer[ServicePerEndpointHttpController](
-      "spe-server",
-      GreeterServicePerEndpointModule,
-      EchoServicePerEndpointModule
-    ),
+  override val server = new EmbeddedHttpServer(
+    twitterServer = new TestHttpServer[ReqRepServicePerEndpointHttpController](
+      "rrspe-server",
+      new GreeterReqRepThriftMethodBuilderClientModule(requestHeaderKey, perRequestTimeoutTunable)),
     args = Seq(
       s"-thrift.clientId=$httpServiceClientId",
       resolverMap(
-        "greeter-thrift-service" -> greeterThriftServer.thriftHostAndPort,
-        "echo-thrift-service" -> echoThriftServer.thriftHostAndPort
-      )
+        "greeter-thrift-service" -> greeterThriftServer.thriftHostAndPort)
     )
   )
 
   override def afterAll() {
     greeterThriftServer.close()
-    echoThriftServer.close()
     super.afterAll()
   }
 
-  test("Greeter.ServicePerEndpoint is available from the injector") {
-    server.injector.instance[Greeter.ServicePerEndpoint] should not be null
+  test("Greeter.ReqRepServicePerEndpoint is available from the injector") {
+    server.injector.instance[Greeter.ReqRepServicePerEndpoint] should not be null
   }
 
   test("Greeter.MethodPerEndpoint is available from the injector") {
     server.injector.instance[Greeter.MethodPerEndpoint] should not be null
-  }
-
-  test("EchoService.ServicePerEndpoint is available from the injector") {
-    server.injector.instance[EchoService.ServicePerEndpoint] should not be null
-  }
-
-  test("EchoService.MethodPerEndpoint is available from the injector") {
-    server.injector.instance[EchoService.MethodPerEndpoint] should not be null
   }
 
   test("Say hi") {
@@ -72,8 +62,6 @@ class DoEverythingServicePerEndpointModuleFeatureTest extends FeatureTest with H
     intercept[Exception] {
       server.httpGet(path = "/hi?name=Bob", andExpect = Ok, withBody = "Hi Bob")
     }
-    // http server route stat
-    server.assertStat("route/hi/GET/response_size", Seq(5))
 
     // per-method -- all the requests in this test were to the same method
     /* assert counters added by ThriftServiceIface#statsFilter */
@@ -89,9 +77,10 @@ class DoEverythingServicePerEndpointModuleFeatureTest extends FeatureTest with H
   }
 
   test("Say hello") {
-    server.httpGet(path = "/hello?name=Bob", andExpect = Ok, withBody = "Hello Bob")
-    // http server route stat
-    server.assertStat("route/hello/GET/response_size", Seq(9))
+    val response =
+      server.httpGetJson[HelloResponse](path = "/hello?name=Bob", andExpect = Ok)
+    assert(response.value == "Hello Bob")
+    assert(!response.headers.empty)
 
     // per-method -- all the requests in this test were to the same method
     /* assert counters added by ThriftServiceIface#statsFilter */
@@ -112,8 +101,6 @@ class DoEverythingServicePerEndpointModuleFeatureTest extends FeatureTest with H
       andExpect = Ok,
       withBody = "Bye Bob of 18 years!"
     )
-    // http server route stat
-    server.assertStat("route/bye/GET/response_size", Seq(20))
 
     // per-method -- all the requests in this test were to the same method
     /* assert counters added by ThriftServiceIface#statsFilter */
@@ -126,38 +113,5 @@ class DoEverythingServicePerEndpointModuleFeatureTest extends FeatureTest with H
     /* assert MethodBuilder counters */
     server.assertCounter("clnt/greeter-thrift-client/bye/logical/requests", 1)
     server.assertCounter("clnt/greeter-thrift-client/bye/logical/success", 1)
-  }
-
-  test("echo 3 times") {
-    server.httpPost(
-      path = "/config?timesToEcho=2",
-      postBody = "",
-      andExpect = Ok,
-      withBody = "2"
-    )
-
-    server.httpPost(
-      path = "/config?timesToEcho=3",
-      postBody = "",
-      andExpect = Ok,
-      withBody = "3"
-    )
-
-    server.httpGet(path = "/echo?msg=Bob", andExpect = Ok, withBody = "BobBobBob")
-    // http server route stats
-    server.assertStat("route/config/POST/response_size", Seq(1, 1))
-    server.assertStat("route/echo/GET/response_size", Seq(9))
-
-    // per-method -- all the requests in this test were to the same method
-    /* assert counters added by ThriftServiceIface#statsFilter */
-    server.assertCounter("clnt/echo-thrift-client/EchoService/echo/requests", 1)
-    server.assertCounter("clnt/echo-thrift-client/EchoService/echo/success", 1)
-    server.assertCounter("clnt/echo-thrift-client/EchoService/echo/failures", 0)
-    /* assert MethodBuilder stats exist */
-    server.getStat("clnt/echo-thrift-client/echo/logical/request_latency_ms") should not be Seq()
-    server.getStat("clnt/echo-thrift-client/echo/retries") should be(Seq(0.0))
-    /* assert MethodBuilder counters */
-    server.assertCounter("clnt/echo-thrift-client/echo/logical/requests", 1)
-    server.assertCounter("clnt/echo-thrift-client/echo/logical/success", 1)
   }
 }