+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/TestThriftServer.scala
@@ -1,5 +1,6 @@
 package com.twitter.inject.thrift.integration
 
+import com.twitter.conversions.time._
 import com.twitter.finagle.{ListeningServer, ThriftMux}
 import com.twitter.inject.server.{PortUtils, TwitterServer}
 import com.twitter.scrooge.ThriftService
@@ -7,6 +8,12 @@ import com.twitter.util.Await
 
 class TestThriftServer(service: ThriftService) extends TwitterServer {
   private val thriftPortFlag = flag("thrift.port", ":0", "External Thrift server port")
+  private val thriftShutdownTimeout = flag(
+    "thrift.shutdown.time",
+    1.minute,
+    "Maximum amount of time to wait for pending requests to complete on shutdown"
+  )
+
   /* Private Mutable State */
   private var thriftServer: ListeningServer = _
 
@@ -18,9 +25,9 @@ class TestThriftServer(service: ThriftService) extends TwitterServer {
   }
 
   onExit {
-    Await.result(thriftServer.close())
+    Await.result(thriftServer.close(thriftShutdownTimeout().fromNow))
   }
 
   /* Overrides */
-  override def thriftPort = Option(thriftServer) map PortUtils.getPort
+  override def thriftPort: Option[Int] = Option(thriftServer) map PortUtils.getPort
 }